feat: embedding API, real weak-ref WeakMap/WeakSet, arbitrary-precision BigInt
Address three review findings:

- Embedding API: Vm fields are now pub(crate) so the public API
  surface is explicit. New public methods define_global/get_global let
  embedders expose host values and read JS globals; new_native_function
  and NativeFn (now pub-used) allow registering native callbacks.
  lib.rs re-exports NativeFn, GcIdx, HeapObj, Error, ErrorKind and keeps
  builtins internal.

- WeakMap/WeakSet: replaced the strong-ref shim with real ephemeron
  semantics. New WeakMapData/WeakSetData hold object keys as raw heap
  indices; GC sweep drops entries whose key was collected, and the trace
  marks a WeakMap value only when its key is already marked. Non-object
  keys throw TypeError, matching spec.

- BigInt: switched Value::BigInt from i128 to num-bigint::BigInt, so
  literals and arithmetic are arbitrary precision (2n**200n is exact).
  All arithmetic/comparison/equality paths updated. Adds num-bigint,
  num-traits, num-integer dependencies.

All existing tests pass; fmt clean.

+++ src/ast.rs
@@ -1,3 +1,4 @@
+use num_bigint::BigInt;
 use std::rc::Rc;
 
 #[derive(Debug, Clone, PartialEq)]
@@ -29,7 +30,7 @@ pub struct ClassMethod {
 #[derive(Debug, Clone, PartialEq)]
 pub enum Expr {
     Number(f64),
-    BigInt(i128),
+    BigInt(BigInt),
     String(Rc<str>),
     TemplateStr(Rc<str>),
     /// Tagged template: `tag`...`` — calls tag(strings, raw, ...exprs).
